examples: add failing test with includes
This adds a test trying to build a requirements txt file simply
containing a `-e other_file.txt`.
We (ab)use the existing flake8.txt file for that.

pypi2nix fails with
> Could not open requirements file: [Errno 2] No such file or directory: '/run/user/1000/pypi2nix/8db5a60850956d906d64b8efa2f5cab0/flake8.txt'
as the referenced file is not copied into the rundir.

related to https://github.com/garbas/pypi2nix/issues/156